[0] warmup/pyproject.toml
[project]
name = "warmup"
version = "0.1.0"
description = ""
authors = [
    {name = "William Correa",email = "wilcorrea@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "openai (>=2.6.1,<3.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
